requirements.txt
streamlit 
PyPDF2 
langchain-community 
langchain 
pillow 
PyMuPDF 
chromadb
llama3.1
nomic-embed-text

AIzaSyCrh708yYBMm3uRRJEZA8H2LFTON5AhpV0